+++ CHANGELOG.md
@@ -7,6 +7,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2021-11-11
+
+### Added
+
+- Added a method to activate an entry within the product catalog.
+
+### Changed
+
+- Changed the ProductCatalog to ignore products that don't match schema.
+
 ## [2.1.1] - 2021-11-10
 
 ### Added